Move test helpers into a more accessible directory
The FuzzTestBuilder previously used for testing bug detectors has
turned out to be quite useful for other tests. Here we move it into
the "tests/" directory so that it can be used by all tests who wish to
run fuzz tests in dedicated processes.
