+++ tests/bug-detectors/command-injection.test.js
@@ -19,7 +19,7 @@ const {
 	FuzzTestBuilder,
 	FuzzingExitCode,
 	// eslint-disable-next-line @typescript-eslint/no-var-requires
-} = require("./helpers.js");
+} = require("../helpers.js");
 // eslint-disable-next-line @typescript-eslint/no-var-requires
 const path = require("path");
 // eslint-disable-next-line @typescript-eslint/no-var-requires
