+++ tests/bug-detectors/path-traversal/fuzz.js
@@ -23,7 +23,7 @@ const fs = require("fs");
 const fsp = require("fs").promises;
 const path = require("path");
 
-const { makeFnCalledOnce } = require("../helpers");
+const { makeFnCalledOnce } = require("../../helpers");
 
 const evil_path = "../../jaz_zer/";
 const safe_path = "../../safe_path/";
